libnetmd: Move send_track() from CLI to netmd_send_track()
